+++ Makefile
@@ -83,7 +83,8 @@ SRCS += hw/acpi/acpi.c
 
 # core
 #SRCS += core/bootrom.c
-SRCS += core/sw_load.c
+SRCS += core/sw_load_common.c
+SRCS += core/sw_load_bzimage.c
 SRCS += core/smbiostbl.c
 SRCS += core/mevent.c
 SRCS += core/gc.c
