+++ include/vmmapi.h
@@ -141,9 +141,4 @@ int	vm_set_ptdev_intx_info(struct vmctx *ctx, uint16_t virt_bdf,
 int	vm_reset_ptdev_intx_info(struct vmctx *ctx, int virt_pin, bool pic_pin);
 
 int	vm_create_vcpu(struct vmctx *ctx, int vcpu_id);
-
-int acrn_parse_kernel(char *arg);
-int acrn_parse_ramdisk(char *arg);
-int acrn_parse_bootargs(char *arg);
-int acrn_sw_load(struct vmctx *ctx);
 #endif	/* _VMMAPI_H_ */
